combine graph with symbols
